+++ test/fixtures/SystemWebViewClient.no-method.java
@@ -1,0 +1,10 @@
+package org.apache.cordova.engine;
+
+import android.webkit.WebView;
+
+public class SystemWebViewClient {
+    @Override
+    public void onPageFinished(WebView view, String url) {
+        super.onPageFinished(view, url);
+    }
+}